Fix ExperimentalBenchmark and FlatSharpEndToEndTests on Linux and macOS
Using properties and items makes sure that paths with backslashes are properly interpreted as paths on Linux and macOS. Using `\` directly inside the Exec task would produce the following error:
> FlatSharpEndToEndTests.csproj(44, 5): [MSB3073] The command "dotnet ..\..\FlatSharp.Compiler\bin\Debug\net5.0\FlatSharp.Compiler.dll --nullable-warnings true -i Grpc\Grpc.fbs -o Grpc\" exited with code 1.

Also, the flatc binaries must have executable permissions set on Linux and macOS else we get this exception:
> System.ComponentModel.Win32Exception (13): Permission denied

Finally, the FBS target must be run before the CoreCompile target instead of the BeforeBuild target to ensure that FlatSharp.Compiler.dll has been built. You see this error when doing a fresh checkout of the project when  the `FlatSharp.Compiler` project has not been built yet.